started fixing java and sql type mistmatching

+++ backend/core/src/main/java/decidish/com/core/model/recipes/IngredientGroup.java
@@ -5,7 +5,7 @@ import java.util.List;
  * Represents one Ingredient required by the recipes (e.g., "Flour - 500g")
  */
 public record IngredientGroup(
-    Long ingredientId,
+    Integer ingredientId,
     String ingredientName,
     Double totalAmountNeeded, // The sum of all recipes (e.g. 500.0)
     // String unit,              // e.g. "g"
